!UPSTREAM: i195: Enable modification of CLOS via debugfs (WIP)
By default the graphics device is configured to use as much of cache as
possible. The cache used by the device does not appear in the "Bitmask
of Shareable Resource with Other executing entities" found in EBX during
enumeration of RDT/CAT. There is thus currently no way to detect or
configure the cache usage between CAT and i915 to ensure expected cache
boundaries are respected.

The existing i915_cache_sharing debugfs file enables user to modify the
active CLOS for the non-render targets on GEN6 and GEN7 hardware. This
feature is supported on later hardware so we enable this modification
for GEN9 also (FIXME: how to enable for GEN6+?) in order to test on APL.

In addition, we add the ability to set the same CLOS not just for
non-render targets, but for render targets also.

When a user thus writes the minimum CLOS (3) to this debugfs file, the
cache usage for both render and non-render targets will be restricted to
the top 1/16th way of the cache (for 16 way cache).

FIXME: commit is only created for testing on APL. We need to work with
graphics folks on the right solution.

Upstreamable solution is being tracked in:
 https://jira01.devtools.intel.com/browse/RT-144

Not-signed-off-by: Reinette Chatre <reinette.chatre@intel.com>