ipipe: add cpuidle control interface
Add a kernel interface for sharing CPU idling control between the host
kernel and a co-kernel. The former invokes ipipe_cpuidle_control()
which the latter should implement, for determining whether entering a
sleep state is ok. This hook should return boolean true if so.

The co-kernel may veto such entry if need be, in order to prevent
latency spikes, as exiting sleep states might be costly depending on
the CPU idling operation being used.