INTERNAL drm/i915: Add plane alpha blending support, v2.
Add plane alpha blending support with the different blend modes.
This has been tested on a icl to show the correct results,
on earlier platforms small rounding errors cause issues. But this
already happens case with fully transparant or fully opaque RGB8888
fb's.

The recommended HW workaround is to disable alpha blending when the
plane alpha is 0 (transparant, hide plane) or 0xff (opaque, disable blending).
This is easy to implement on any platform, so just do that.

The tests for userspace are also available, and pass on gen11.

Changes since v1:
- Change mistaken < 0xff0 to 0xff00.
- Only set PLANE_KEYMSK_ALPHA_ENABLE when plane alpha < 0xff00, ignore blend mode.
- Rework disabling FBC when per pixel alpha is used.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
[mlankhorst: Change MISSING_CASE default to explicit alpha disable (mattrope)]
Link: https://patchwork.freedesktop.org/patch/244642/
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>

v2:
- Rebase into lts/4.19/base branch
- Since GP2.0 is using restrict_planes
- Add alpha property and blend mode property in intel_skl_plane_create
- which func is used for intel_crtc_init_restrict_planes
- Removing calls from original patch (UPSTREAM)

Tacked-On: OLINUX-3689
Signed-off-by: kimsehun <se.hun.kim@intel.com>