staging: Add AVnu based Intel IGB driver
This initial commit adds the AVnu based Intel IGB driver
for the I210 ethernet controller to drivers/staging. It
mirrors and will remain in sync with what's hosted at
https://github.com/AVnu/igb_avb/tree/master/kmod which is
the github repo for the AVnu IGB driver.

It can be built as module with CONFIG_IGB_AVB=m. This is
*exclusive* with the in-tree IGB driver located at drivers/
net/ethernet/intel/igb/, so, if one is enabled, then the
other one will get disabled and vice-versa.

Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>