x86/process: ipipe: prevent false positive in testing for preemptible section
We may call switch_to_bitmap() from the head stage when a co-kernel
switches to a new thread context, in which case IRQs are hard
disabled, although the root stage might be unstalled.

Make sure the preemption debug code in refresh_tss_limit() is aware of
this.

Analogously to refresh_tss_limit, we need to enhance the debug check
in invalidate_tss_limit so that KVM's __vmx_load_host_state can be
called from __ipipe_handle_vm_preemption.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>