media: Set IPU device name
The ipu device name in debugfs is unfixed and with
bad readability, and the bad thing is that the changing
default ipu device name make pnp's application program
base on it couldn't execute correctly, as the program
couldn't find the target file in debugfs once the
ipu device name changed.
The original sysfs path include bus number as default
device name looks like:
"/sys/kernel/debug/0000:02:05.0/",
And then the path fix to "/sys/kernel/debug/intel-ipu/"
with the change.

Change-Id: Ie4d080ef011bb0eb8688d833a0cd3167e14cbf4c
Tracked-On: PKT-1691
Signed-off-by: Haiping Yang <haipingx.yang@intel.com>
Signed-off-by: Meng Wei <wei.meng@intel.com>