Kernel/VHM: VM_list changes the lock from mutex to rw_lock to avoid sleeping
VM_list records all the added vm for each DM. And it is accessed by using
mutex lock. As it is accessed in created/removed and softireq tasklet, the
spinlock should be used instead. And as the read is used frequently, the
rwlock is used(similar to spinlock except it allows multiple readers).
And when the vhm_dev_release is closed, the vhm_vm will be removed from
the list and its memory is released only when its refcnt is decreased to
zero.

Tracked-On: projectacrn/acrn-hypervisor#2085
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Shuo Liu <shuo.a.liu@intel.com>
Reviewed-by: Yin FengWei <fengwei.yin@intel.com>
Tracked-On: PKT-1617