ASoC: Intel: Skl: Virt: Add static map of domain_ids
If an ALSA control doesn't belong to any widget, BE is unable to check
access rights of that control due to unassigned domain_id. The change
provides static, hardcoded map between control and domain_id to fix
this issue. This is just a temporary workaround. The permanent solution
requires code refactoring

Change-Id: Ic92bc8276b30e0f7fa72e6c211d3ed7804c23b77
Tracked-On: OAM-76301
Reviewed-by: Lewandowski, Gustaw <gustaw.lewandowski@intel.com>
Tested-by: Lewandowski, Gustaw <gustaw.lewandowski@intel.com>