ipipe: Disable rcuidle trace path when running over the head domain
We do not need the special handling of __DO_TRACE(..., rcuidle=1) when
running over the head domain. In fact, we cannot use it because it
switches to srcu which is incompatible with that context. It's safe to
switch to normal RCU because no head domain caller of a trace_*_rcuidle
tracepoints should do this from rcu-problematic paths, specifically
idle.

Ported from the dovetail queue.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>