ipipe: use tsc_khz to instead of cpu_khz when cpu support tsc feature
For some x86 cpu platforms, tsc frequency will be refined calibration
with system boot, which will lead to tsc_khz is different with cpu_khz,
but the right tsc frequency should be tsc_khz which is registed to tsc clocksource.
as xenomai still use cpu_khz as tsc clocksource, which will lead to real time of cycle is not correct.

Signed-off-by: Zhang, Wei E <wei.e.zhang@intel.com>