ipipe: pinctrl: enable basic pipeline support
This patch enables the Intel pinctrl/GPIO core driver to operate in a
pipelined interrupt system. However, it does not allow chained GPIO
IRQs to be handled from the head stage of such pipeline yet. In other
words, the chained GPIO interrupts can safely be handled from the
in-band stage when CONFIG_IPIPE is turned on, but cannot be routed to
a real-time application.

Enabling full support will require the I-pipe core to handle IRQs
chained from a shared parent interrupt natively, which it is not
implemented at the moment.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>